requirements_vertex_ai.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
python-docx==1.1.2
google-cloud-aiplatform>=1.38.0
supabase==2.6.0
httpx==0.27.2
Pillow==10.4.0
